+++ src/hooks/useSettings.tsx
@@ -1,0 +1,7 @@
+import { useContext } from "react";
+
+import { SettingsContext } from "@/providers/settings.provider";
+
+const useSettings = () => useContext(SettingsContext);
+
+export default useSettings;