+++ src/providers/index.tsx
@@ -1,9 +1,10 @@
-import { Geiger } from 'react-geiger';
+import { Geiger } from "react-geiger";
 
-import AnalyticsProvider from '@/providers/analytics.provider';
-import PlatformProvider from '@/providers/platform.provider';
-import ThemeProvider from '@/providers/theme.provider';
-import { Notifications } from '@mantine/notifications';
+import AnalyticsProvider from "@/providers/analytics.provider";
+import PlatformProvider from "@/providers/platform.provider";
+import SettingsProvider from "@/providers/settings.provider";
+import ThemeProvider from "@/providers/theme.provider";
+import { Notifications } from "@mantine/notifications";
 
 interface AppProvider {
   children: React.ReactNode;
@@ -12,15 +13,16 @@ interface AppProvider {
 const AppProvider: React.FC<AppProvider> = ({ children }) => {
   return (
     <Geiger renderTimeThreshold={50}>
-      <AnalyticsProvider>
-        <ThemeProvider>
-          <PlatformProvider>
-            
-            <Notifications />
-            {children}
-          </PlatformProvider>
-        </ThemeProvider>
-      </AnalyticsProvider>
+      <SettingsProvider>
+        <AnalyticsProvider>
+          <ThemeProvider>
+            <PlatformProvider>
+              <Notifications />
+              {children}
+            </PlatformProvider>
+          </ThemeProvider>
+        </AnalyticsProvider>
+      </SettingsProvider>
     </Geiger>
   );
 };